+++ internal/cmd/gtrace/main.go
@@ -18,7 +18,7 @@ import (
 	"strings"
 	_ "unsafe" // For go:linkname.
 
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 )
 
 //go:linkname build_goodOSArchFile go/build.(*Context).goodOSArchFile
@@ -287,7 +287,7 @@ func buildFunc(info types.Info, traces map[string]*Trace, fn *ast.FuncType) (ret
 	case *ast.FuncType:
 		result, err := buildFunc(info, traces, x)
 		if err != nil {
-			return nil, errors.WithStackTrace(err)
+			return nil, xerrors.WithStackTrace(err)
 		}
 		ret.Result = append(ret.Result, result)
 		return ret, nil
@@ -405,7 +405,7 @@ func scanBuildConstraints(r io.Reader) (cs []string, err error) {
 	for {
 		line, err := br.ReadBytes('\n')
 		if err != nil {
-			return nil, errors.WithStackTrace(err)
+			return nil, xerrors.WithStackTrace(err)
 		}
 		line = bytes.TrimSpace(line)
 		if comm := bytes.TrimPrefix(line, []byte("//")); !bytes.Equal(comm, line) {