+++ internal/coordination/client.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/coordination"
 	"github.com/ydb-platform/ydb-go-sdk/v3/coordination/config"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/operation"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/scheme"
 )
 
@@ -118,11 +118,11 @@ func (c *client) DescribeNode(
 		},
 	)
 	if err != nil {
-		return nil, nil, errors.WithStackTrace(err)
+		return nil, nil, xerrors.WithStackTrace(err)
 	}
 	err = proto.Unmarshal(response.GetOperation().GetResult().GetValue(), &result)
 	if err != nil {
-		return nil, nil, errors.WithStackTrace(err)
+		return nil, nil, xerrors.WithStackTrace(err)
 	}
 	return scheme.InnerConvertEntry(result.GetSelf()), &coordination.NodeConfig{
 		Path:                     result.GetConfig().GetPath(),