+++ internal/ratelimiter/errors/errors_test.go
@@ -4,14 +4,14 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 )
 
 func TestIsAcquireError(t *testing.T) {
 	for _, err := range []error{
 		&acquireError{},
-		errors.WithStackTrace(&acquireError{}),
-		errors.WithStackTrace(fmt.Errorf("%w", &acquireError{})),
+		xerrors.WithStackTrace(&acquireError{}),
+		xerrors.WithStackTrace(fmt.Errorf("%w", &acquireError{})),
 	} {
 		t.Run("", func(t *testing.T) {
 			if !IsAcquireError(err) {
@@ -24,8 +24,8 @@ func TestIsAcquireError(t *testing.T) {
 func TestToAcquireError(t *testing.T) {
 	for _, err := range []error{
 		&acquireError{},
-		errors.WithStackTrace(&acquireError{}),
-		errors.WithStackTrace(fmt.Errorf("%w", &acquireError{})),
+		xerrors.WithStackTrace(&acquireError{}),
+		xerrors.WithStackTrace(fmt.Errorf("%w", &acquireError{})),
 	} {
 		t.Run("", func(t *testing.T) {
 			ae := ToAcquireError(err)